requirements.txt
ffmpeg-python==0.2.0
future==1.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1